refactor: move gateway api structs into api module
