+++ modules/fedimint-lnv2-client/src/api.rs
@@ -2,6 +2,7 @@ use std::collections::{BTreeMap, BTreeSet};
 use std::time::Duration;
 
 use bitcoin30::secp256k1::schnorr::Signature;
+use bitcoin30::secp256k1::PublicKey;
 use fedimint_api_client::api::{
     FederationApiExt, FederationResult, IModuleFederationApi, PeerResult,
 };
@@ -22,8 +23,9 @@ use fedimint_lnv2_common::ContractId;
 use lightning_invoice::Bolt11Invoice;
 use rand::seq::SliceRandom;
 use serde::{Deserialize, Serialize};
+use thiserror::Error;
 
-use crate::{Bolt11InvoiceDescription, GatewayConnectionError, LightningInvoice, RoutingInfo};
+use crate::{Bolt11InvoiceDescription, LightningInvoice};
 
 const RETRY_DELAY: Duration = Duration::from_secs(1);
 
@@ -203,6 +205,14 @@ pub trait GatewayConnection: std::fmt::Debug {
     ) -> Result<Result<[u8; 32], Signature>, GatewayConnectionError>;
 }
 
+#[derive(Error, Debug, Clone, Eq, PartialEq)]
+pub enum GatewayConnectionError {
+    #[error("The gateway is unreachable: {0}")]
+    Unreachable(String),
+    #[error("The gateway returned an error for this request: {0}")]
+    Request(String),
+}
+
 #[derive(Debug)]
 pub struct RealGatewayConnection;
 
@@ -305,4 +315,82 @@ pub struct SendPaymentPayload {
     pub contract: OutgoingContract,
     pub invoice: LightningInvoice,
     pub auth: Signature,
+}
+
+#[derive(Debug, Clone, Eq, PartialEq, Hash, Serialize, Deserialize)]
+pub struct RoutingInfo {
+    /// The public key of the gateways lightning node. Since this key signs the
+    /// gateways invoices the senders client uses it to differentiate between a
+    /// direct swap between fedimints and a lightning swap.
+    pub lightning_public_key: PublicKey,
+    /// The public key of the gateways client module. This key is used to claim
+    /// or cancel outgoing contracts and refund incoming contracts.
+    pub module_public_key: PublicKey,
+    /// This is the fee the gateway charges for an outgoing payment. The senders
+    /// client will use this fee in case of a direct swap.
+    pub send_fee_minimum: PaymentFee,
+    /// This is the default total fee the gateway recommends for an outgoing
+    /// payment in case of a lightning swap. It accounts for the additional fee
+    /// required to reliably route this payment over lightning.
+    pub send_fee_default: PaymentFee,
+    /// This is the minimum expiration delta in block the gateway requires for
+    /// an outgoing payment. The senders client will use this expiration delta
+    /// in case of a direct swap.
+    pub expiration_delta_minimum: u64,
+    /// This is the default total expiration the gateway recommends for an
+    /// outgoing payment in case of a lightning swap. It accounts for the
+    /// additional expiration delta required to successfully route this payment
+    /// over lightning.
+    pub expiration_delta_default: u64,
+    /// This is the fee the gateway charges for an incoming payment.
+    pub receive_fee: PaymentFee,
+}
+
+impl RoutingInfo {
+    pub fn send_parameters(&self, invoice: &Bolt11Invoice) -> (PaymentFee, u64) {
+        if invoice.recover_payee_pub_key() == self.lightning_public_key {
+            (self.send_fee_minimum.clone(), self.expiration_delta_minimum)
+        } else {
+            (self.send_fee_default.clone(), self.expiration_delta_default)
+        }
+    }
+}
+
+#[derive(Debug, Clone, Eq, PartialEq, PartialOrd, Hash, Serialize, Deserialize)]
+pub struct PaymentFee {
+    pub base: Amount,
+    pub parts_per_million: u64,
+}
+
+impl PaymentFee {
+    /// This is the maximum send fee of one and a half percent plus one hundred
+    /// satoshis a correct gateway may recommend as a default. It accounts for
+    /// the fee required to reliably route this payment over lightning.
+    pub const SEND_FEE_LIMIT: PaymentFee = PaymentFee {
+        base: Amount::from_sats(100),
+        parts_per_million: 15_000,
+    };
+
+    /// This is the maximum receive fee of half of one percent plus fifty
+    /// satoshis a correct gateway may recommend as a default.
+    pub const RECEIVE_FEE_LIMIT: PaymentFee = PaymentFee {
+        base: Amount::from_sats(50),
+        parts_per_million: 5_000,
+    };
+
+    pub fn add_to(&self, msats: u64) -> Amount {
+        Amount::from_msats(msats.saturating_add(self.absolute_fee(msats)))
+    }
+
+    pub fn subtract_from(&self, msats: u64) -> Amount {
+        Amount::from_msats(msats.saturating_sub(self.absolute_fee(msats)))
+    }
+
+    fn absolute_fee(&self, msats: u64) -> u64 {
+        msats
+            .saturating_mul(self.parts_per_million)
+            .saturating_div(1_000_000)
+            .checked_add(self.base.msats)
+            .expect("The division creates sufficient headroom to add the base fee")
+    }
 }
