+++ modules/fedimint-lnv2-tests/tests/mock.rs
@@ -9,10 +9,10 @@ use fedimint_core::secp256k1::KeyPair;
 use fedimint_core::util::SafeUrl;
 use fedimint_core::{apply, async_trait_maybe_send, Amount};
 use fedimint_ln_common::bitcoin;
-use fedimint_lnv2_client::api::GatewayConnection;
-use fedimint_lnv2_client::{
-    Bolt11InvoiceDescription, GatewayConnectionError, LightningInvoice, PaymentFee, RoutingInfo,
+use fedimint_lnv2_client::api::{
+    GatewayConnection, GatewayConnectionError, PaymentFee, RoutingInfo,
 };
+use fedimint_lnv2_client::{Bolt11InvoiceDescription, LightningInvoice};
 use fedimint_lnv2_common::contracts::{IncomingContract, OutgoingContract, PaymentImage};
 use lightning_invoice::{
     Bolt11Invoice, Currency, InvoiceBuilder, PaymentSecret, DEFAULT_EXPIRY_TIME,
